// swiyu-ecosystem-portal-ui/src/app/api/generated/model/identifier-response.ts
/**
 * SWIYU Ecosystem Portal SCS API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import {IdentifierStatus} from './identifier-status';

export interface IdentifierResponse {
  did?: string;
  id?: string;
  description?: string;
  status?: IdentifierStatus;
}
export namespace IdentifierResponse {}