// swiyu-ecosystem-portal-ui/src/app/api/generated/model/identifier-status.ts
/**
 * SWIYU Ecosystem Portal SCS API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export const IdentifierStatus = {
  NotInitialized: 'NOT_INITIALIZED',
  Initialized: 'INITIALIZED',
  UserDeactivated: 'USER_DEACTIVATED',
  DeactivatedByMigrationBecauseOfUnsupportedFormat: 'DEACTIVATED_BY_MIGRATION_BECAUSE_OF_UNSUPPORTED_FORMAT'
} as const;
export type IdentifierStatus = (typeof IdentifierStatus)[keyof typeof IdentifierStatus];